requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
logging>=0.4.0
Pillow>=10.0.0
retry>=0.9.2
schedule>=1.2.0
huggingface_hub>=0.19.0
